FINALLY fixed the many to one mapping. Circular ref removed, variable names making no sense fixed, genius idea to change the ref from a many to one to a one to many stopped in its tracks (not a smart idea after all).
Justification for keeping the ref a many-to-one: one bigger trip is better than many small ones. By keeping the id of the board in the notes we can get all the notes belonging to a board with a single query. Had I reversed the relationship (1:m) I'd have to make n queries to get the notes